Met à jour l'aide au permis `hauts_de_france_aide_permis` et corrige Haut-de-France en
